Moving to a data structure for level data.
By putting it in a data structure populated when the game is initialized 
or reset, we can avoid calling Image.At() to see what object is in what 
position. This is because Image.At() can be costly to perform. That's 
not really going to make a difference when it comes to these super tiny 
level map images, but it could, theoretically. Therefore, using a data 
structure isn't a bad idea.

Extending Level image down a row to fill the screen.